feat: add clipping to lines, update story
